app/src/test/java/com/example/moviesapp/ui/search/SearchMovieViewModelTest.kt
package com.example.moviesapp.ui.search

import com.example.moviesapp.TestStubs
import com.example.moviesapp.TestStubs.Companion.validMovie
import com.example.moviesapp.TestStubs.Companion.validSearchTerm
import com.example.moviesapp.data.repository.MoviesRepositoryImpl
import io.mockk.clearAllMocks
import io.mockk.coEvery
import io.mockk.junit5.MockKExtension
import io.mockk.mockkClass
import kotlinx.coroutines.*
import kotlinx.coroutines.flow.toList
import kotlinx.coroutines.test.*
import org.junit.jupiter.api.AfterEach
import org.junit.jupiter.api.Assertions.*

import org.junit.jupiter.api.BeforeEach
import org.junit.jupiter.api.Test
import org.junit.jupiter.api.parallel.Execution
import org.junit.jupiter.api.parallel.ExecutionMode

@Execution(ExecutionMode.CONCURRENT)
@MockKExtension.ConfirmVerification
@OptIn(ExperimentalCoroutinesApi::class)
internal class SearchMovieViewModelTest {

    private lateinit var viewModel: SearchMovieViewModel
    private lateinit var moviesRepository: MoviesRepositoryImpl

    @OptIn(DelicateCoroutinesApi::class)
    private val mainThreadSurrogate = newSingleThreadContext("UI thread")

    @BeforeEach
    fun setUp() {
        moviesRepository = mockkClass(MoviesRepositoryImpl::class)
        viewModel = SearchMovieViewModel(moviesRepository)
        Dispatchers.setMain(mainThreadSurrogate)
    }

    @AfterEach
    fun finish() {
        clearAllMocks()
        Dispatchers.resetMain() // reset the main dispatcher to the original Main dispatcher
        mainThreadSurrogate.close()
    }

    @Test
    fun `test SearchMovieViewModel initSearch using valid searchTerm returns state with movies`() = runBlockingTest {

        coEvery {
            moviesRepository.searchMovieUsingQuery(validSearchTerm)
        } returns TestStubs.validMovieResponse.listOfMovies

        val results = mutableListOf<SearchMovieUIState>()

        val job = launch {
            viewModel.searchMovieUIState.toList(results)
        }

        viewModel.initSearch(validSearchTerm)

        assertEquals(
            SearchMovieUIState(movies = null, isLoading = false, isError = false),
            results[0],
            "initial state of SearchMovieUIState"
        )

        assertEquals(
            SearchMovieUIState(movies = null, isLoading = true, isError = false),
            results[1],
            "loading state of SearchMovieUIState"
        )

        // TODO: issue with results[2], can't find a way to block execution till
        //  viewModel.initSearch finishes
//        assertEquals(
//            SearchMovieUIState(movies = TestStubs.validMovieResponse.listOfMovies, isLoading = true, isError = false),
//            results[2]
//        )

        job.cancel()
    }

    @Test
    fun `test SearchMovieViewModel initSearch using empty string returns state with empty list of movies`() = runBlockingTest {
        val results = mutableListOf<SearchMovieUIState>()

        val job = launch {
            viewModel.searchMovieUIState.toList(results)
        }

        viewModel.initSearch("")

        assertEquals(
            SearchMovieUIState(movies = null, isLoading = false, isError = false),
            results[0],
            "initial state of SearchMovieUIState"
        )

        assertEquals(
            SearchMovieUIState(movies = null, isLoading = true, isError = false),
            results[1],
            "loading state of SearchMovieUIState"
        )

        // TODO: this test is flaky, figure out why
//        assertEquals(
//            SearchMovieUIState(movies = emptyList(), isLoading = false, isError = false),
//            viewModel.searchMovieUIState.value,
//            "empty list state of SearchMovieUIState"
//        )

        job.cancel()
    }

    // TODO: because of the above issues manually changed some methods from private to internal
    //  to be able to test them without testing initSearch directly

    @Test
    fun `test SearchMovieViewModel handleLoading updates searchMovieUIState correctly`() {
        viewModel.handleLoading()
        assertEquals(
            SearchMovieUIState(movies = null, isLoading = true, isError = false),
            viewModel.searchMovieUIState.value
        )
    }

    @Test
    fun `test SearchMovieViewModel handleMovies updates searchMovieUIState correctly`() {
        viewModel.handleMovies(listOf(validMovie))
        assertEquals(
            SearchMovieUIState(movies = listOf(validMovie), isLoading = false, isError = false),
            viewModel.searchMovieUIState.value
        )
    }

    @Test
    fun `test SearchMovieViewModel handleError updates searchMovieUIState correctly`() {
        viewModel.handleError()
        assertEquals(
            SearchMovieUIState(movies = null, isLoading = false, isError = true),
            viewModel.searchMovieUIState.value
        )
    }

}